Gracefully handle project errors
