+++ inlang/source-code/paraglide/paraglide-js/src/index.ts
@@ -1,6 +1,7 @@
 export { compile } from "./compiler/compile.js"
 export { writeOutput } from "./services/file-handling/write-output.js"
 export { Logger, type LoggerOptions } from "./services/logger/index.js"
+export { classifyProjectErrors } from "./services/error-handling.js"
 
 export type MessageIndexFunction<T extends string> = (
 	params: Record<string, never>,
